requirements.txt
anyio==3.6.2
certifi==2022.12.7
cffi==1.15.1
cryptography==40.0.2
decorator==5.1.1
future==0.18.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pycparser==2.21
python-telegram-bot==12.8
rfc3986==1.5.0
sniffio==1.3.0
tornado==6.3.1
